+++ csobclient/http/__init__.py
@@ -1,0 +1,16 @@
+"""Package for dealing with HTTP."""
+from .base import (
+    HTTPClient,
+    HTTPConnectionError,
+    HTTPRequestError,
+    HTTPResponse,
+    HTTPTimeoutError,
+)
+
+__all__ = [
+    "HTTPClient",
+    "HTTPConnectionError",
+    "HTTPRequestError",
+    "HTTPResponse",
+    "HTTPTimeoutError",
+]