+++ csobclient/v19/http/__init__.py
@@ -1,9 +1,0 @@
-"""Module responsible for HTTP requests."""
-from .client import (
-    HTTPClient,
-    HTTPRequestError,
-    HTTPConnectionError,
-    HTTPTimeoutError,
-    HTTPResponse,
-)
-from .requests_client import RequestsHTTPClient