+++ tests/http/base/__init__.py
@@ -1,0 +1,1 @@
+"""Test for the base module."""